pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "hypergraph"
version = "0.1.0"
description = "Uses Hypergraphs to Simulate Drug Combinations for TB."
authors = ["Logan Reed <me@loganreed.org>"]

[tool.poetry.dependencies]
python = ">=3.8,<4"
pandas = "^1.4.1"
openpyxl = "^3.0.9"
biopython = "^1.79"
reportlab = "^3.6.9"
GEOparse = "^2.0.3"

[tool.poetry.dev-dependencies]
tox = "^3.24.5"
pytest = "^7.0"
black = "^22.3.0"
